+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 
 ## vN.N.N
+* Add eme-form-block to Blocksmith.
 * cms crash to white screen after paste